Create new operation type (#13)
* Refactor code into separate files.
* Add methods for creating new op type and code objects.
* Update push method so that uses op type and user ids from Aquarium instance.
