+++ pxfish/operation_types.py
@@ -1,0 +1,7 @@
+"""
+Utilities for operation types.
+"""
+
+
+def operation_type_code_names():
+    return ['protocol', 'precondition', 'cost_model', 'documentation', 'test']
